@bkamins 2020-09-20
Adding [BREAKING] as it changes return value order when `sort=false` (minor but still...).