@bkamins 2020-09-20
Looks good thank you - the open comments are minor. I am OK to merge the PR when you resolve them.