@nalimilan 2020-09-21
I've added a commit to use the generic fallback when the pools are not unique, plus a few small things.