@nalimilan 2020-09-21
> Additionally - a general question - I hope that the tests cover all code paths (in particular the path when we have too many groups so we should invoke a slow path).

Yes, that's covered by the new tests (there was already one but quite limited). We don't have an example of an array for which `refpool` contains duplicates, but since the path is the same it should be OK.