@bkamins 2020-10-11
So - I understand we merge it once CI passes. Right?